refactor code input_handler fixed bug can't get negative number, add more function but show coming soon...

+++ src/input_handler.h
@@ -5,7 +5,7 @@
 #ifndef INPUT_HANDLER_H
 #define INPUT_HANDLER_H
 
-size_t get_mat_cnt();
+int get_mat_cnt();
 Matrix* get_mat(int cnt);
 
 #endif
